requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
bs4==0.0.2
fastapi==0.115.14
idna==3.10
lxml==5.4.0
numpy==2.3.1
pandas==2.3.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-multipart==0.0.20
pytz==2025.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uvicorn[standard]==0.30.6
